+++ src/app/app.routes.ts
@@ -4,13 +4,14 @@ import { AdminLayout } from './features/auth/layouts/admin-layout/admin-layout';
 import { ClientLayout } from './features/auth/layouts/client-layout/client-layout';
 import { Intranet } from './features/admin/intranet/intranet';
 import { authGuard } from './core/guards/auth-guard';
-import { Nosotros } from './pages/nosotros/nosotros';
+import { Nosotros } from './features/publico/nosotros/nosotros';
+import { Error } from './features/publico/error/error';
 
 export const routes: Routes = [
     {
-        path: 'admin', component: AdminLayout, canActivate:[authGuard], children: [
+        path: 'admin', component: AdminLayout, canActivate: [authGuard], children: [
             {
-                path:'intranet', component:Intranet, title:'Intranet'
+                path: 'intranet', component: Intranet, title: 'Intranet'
             }
         ]
     },
@@ -19,12 +20,13 @@ export const routes: Routes = [
             {
                 path: 'login', component: Login, title: "Iniciar Sesión"
             },
-            {path:'nosotros', component:Nosotros, title:"Nosotros"}
+            { 
+                path: 'nosotros', component: Nosotros, title: "Nosotros" 
+            },
+            {
+                path: '**', component: Error, title: "Pagina de error"
+            }
         ]
-    },
-
-    
-
-  
+    }
 
 ];
