+++ src/app/features/publico/error/error.html
@@ -1,0 +1,1 @@
+<p>error works!</p>
