+++ src/app/features/publico/error/error.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-error',
+  imports: [],
+  templateUrl: './error.html',
+  styleUrl: './error.scss'
+})
+export class Error {
+
+}
