+++ src/app/features/publico/nosotros/nosotros.html
@@ -1,0 +1,86 @@
+<!-- HEADER COMPLETO A LO ANCHO -->
+<header class="hero">
+    <div class="hero__container">
+        <div class="hero__content">
+            <h1 class="hero__title">ROYAL SUITES HOTEL</h1>
+            <p class="hero__subtitle">
+                Bienvenido a un lugar donde el descanso, la calidez y el servicio personalizado se unen para ofrecerte
+                una
+                experiencia de hospedaje inolvidable
+            </p>
+        </div>
+        <img class="hero__img" src="img/1Nosotros.png" alt="Hotel fachada">
+    </div>
+</header>
+
+<!-- ACERCA DE NOSOTROS CON FONDO A LO ANCHO -->
+<section class="about">
+    <div class="about__container">
+        <h2>Acerca de nosotros</h2>
+        <div class="about__content">
+            <img src="img/2Nosotros.png" alt="Acerca de nosotros" style="width: 290px; height: 190px;">
+            <p>
+                Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the
+                industry's
+                standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it
+                to make a type specimen book. It has survived not only five centuries,standard dummy text ever since the
+                1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It
+                has survived not only five centuries,
+            </p>
+        </div>
+    </div>
+</section>
+
+<!-- MISIÓN -->
+<section class="mission">
+    <div class="mission__container">
+        <h2>Misión</h2>
+        <div class="mission__content">
+            <img src="img/3Nosotros.png" alt="Recepción del hotel" style="width: 290px; height: 190px;" >
+            <p>Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the
+                industry's
+                standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it
+                to make a type specimen book. It has survived not only five centuries,standard dummy text ever since the
+                1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It
+                has survived not only five centuries,
+            </p>
+        </div>
+    </div>
+</section>
+
+<!-- VISIÓN CON FONDO -->
+<section class="vision">
+    <div class="vision__container">
+        <h2>Visión</h2>
+        <div class="vision__content">
+            <img src="img/4Nosotros.png" alt="Lobby moderno" style="width: 290px; height: 190px;">
+            <p>
+                Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the
+                industry's
+                standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it
+                to make a type specimen book. It has survived not only five centuries,standard dummy text ever since the
+                1500s, when an unknown printer took a galley of type and scrambled it to make a type specimen book. It
+                has survived not only five centuries,
+            </p>
+        </div>
+    </div>
+</section>
+
+<!-- RESEÑAS -->
+<section class="reviews">
+    <div class="review">
+        <div class="review__stars">★★★★★</div>
+        <p class="review__text">Desde el momento en que llegamos fue una excelente experiencia</p>
+        <div class="review__author">Laura G. – Lima, Perú</div>
+    </div>
+    <div class="review">
+        <div class="review__stars">★★★★</div>
+        <p class="review__text">Desde el momento en que llegamos</p>
+        <div class="review__author">Laura G. – Lima, Perú</div>
+    </div>
+    <div class="review">
+        <div class="review__stars">★★★</div>
+        <p class="review__text">Desde el momento en que llegamos</p>
+        <div class="review__author">Laura G. – Lima, Perú</div>
+    </div>
+</section>
