+++ src/app/features/publico/nosotros/nosotros.ts
@@ -1,0 +1,11 @@
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-nosotros',
+  imports: [],
+  templateUrl: './nosotros.html',
+  styleUrl: './nosotros.scss'
+})
+export class Nosotros {
+
+}
