--- src/app/pages/nosotros/nosotros.scss
@@ -1,165 +1,0 @@
-$primary-color: #ffb300;
-$secondary-color: #162550;
-$background-light: #f7f7fc;
-$text-dark: #222;
-$text-light: #fff;
-
-// Hero
-.hero {
-  width: 100%;
-  padding: 2rem 1rem;
-  margin-bottom: 18rem;
-
-  .hero__container {
-    display: flex;
-    flex-direction: row;
-    justify-content: space-between;
-    align-items: center;
-    flex-wrap: wrap;
-    max-width: 1100px;
-    margin: 0 auto;
-    
-  }
-
-  &__content {
-    flex: 1;
-    text-align: left;
-    padding-right: 1rem;
-    min-width: 280px;
-  }
-
-  &__title {
-    color: $primary-color;
-    font-size: 2.2rem;
-    font-weight: bold;
-    margin-bottom: 0.5rem;
-  }
-
-  &__subtitle {
-    font-size: 1.1rem;
-    color: $text-dark;
-    font-weight: 500;
-  }
-
-  &__img {
-    flex: 1;
-    max-width: 400px;
-    width: 100%;
-    border-radius: 0.5rem;
-  }
-}
-
-section {
-  width: 100%;
-
-  > div {
-    max-width: 900px;
-    margin: 0 auto;
-    padding: 2rem 1rem;
-  }
-
-  h2 {
-    color: $secondary-color;
-    text-align: center;
-    font-size: 2.2rem;
-    margin-bottom: 1.2rem;
-  }
-}
-
-.about {
-  background-color: #C2C3E8;
-}
-.vision {
-  background-color: #C2C3E8;
-}
-
-// Layout de contenido
-.about__content,
-.mission__content,
-.vision__content {
-  display: flex;
-  align-items: center;
-  gap: 2rem;
-
-  img {
-    width: 230px;
-    border-radius: 0.5rem;
-    object-fit: cover;
-  }
-
-  p {
-    flex: 1;
-    font-size: 1rem;
-    color: $text-dark;
-  }
-}
-
-// Orden especial
-.mission__content {
-  flex-direction: row-reverse;
-}
-.vision__content {
-  flex-direction: row;
-}
-
-// Reseñas
-.reviews {
-  background: $background-light;
-  border-radius: 0.7rem;
-  padding: 2rem 1rem;
-
-  .review {
-    background: $text-light;
-    border: 1px solid #ececec;
-    border-radius: 0.5rem;
-    padding: 1rem 1.2rem;
-    margin-bottom: 1.2rem;
-    box-shadow: 0 2px 6px rgba(90, 90, 158, 0.04);
-
-    &__stars {
-      color: $primary-color;
-      font-size: 1.2rem;
-      margin-bottom: 0.5rem;
-    }
-
-    &__text {
-      font-size: 1rem;
-      margin-bottom: 0.7rem;
-    }
-
-    &__author {
-      font-size: 0.95rem;
-      color: $secondary-color;
-      font-weight: 500;
-    }
-  }
-}
-
-// Responsividad
-@media (max-width: 750px) {
-  .hero {
-    margin-bottom: 30rem;
-    .hero__container {
-      flex-direction: column;
-      text-align: center;
-    }
-
-    &__img {
-      margin-top: 1rem;
-      max-width: 350px;
-    }
-  }
-
-  .about__content,
-  .mission__content,
-  .vision__content {
-    flex-direction: column !important;
-    text-align: center;
-
-    img {
-      width: 100%;
-      max-width: 320px;
-      margin-bottom: 1rem;
-    }
-  }
-}
