--- src/app/pages/nosotros/nosotros.ts
@@ -1,11 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-nosotros',
-  imports: [],
-  templateUrl: './nosotros.html',
-  styleUrl: './nosotros.scss'
-})
-export class Nosotros {
-
-}
